requirements.txt
Django==2.2
djangorestframework==3.12.4
psycopg2-binary==2.8.6
django-environ==0.4.5
django-filter==2.4.0
